Post-$&newparse changes and cleanup
Not all of this actually needs $&newparse as a prerequisite, like
input->eof and removing the various fill functions.
