+++ heredoc.c
@@ -40,11 +40,10 @@ extern Tree *snarfheredoc(Parser *p, const char *eof, Boolean quoted) {
 		yyerror(p, "here document eof-marker contains a newline");
 		return NULL;
 	}
-	p->input->ignoreeof = TRUE;
 
 	for (tree = NULL, tailp = &tree, buf = openbuffer(0);;) {
 		int c;
-		print_prompt2(p);
+		p->input->lineno++;
 		for (s = (unsigned char *) eof; (c = get(p)) == *s; s++)
 			;
 		if (*s == '\0' && (c == '\n' || c == EOF)) {
@@ -60,7 +59,7 @@ extern Tree *snarfheredoc(Parser *p, const char *eof, Boolean quoted) {
 			if (c == EOF) {
 				yyerror(p, "incomplete here document");
 				freebuffer(buf);
-				p->input->ignoreeof = FALSE;
+				p->input->eof = FALSE;
 				return NULL;
 			}
 			if (c == '$' && !quoted && (c = get(p)) != '$') {
@@ -75,7 +74,7 @@ extern Tree *snarfheredoc(Parser *p, const char *eof, Boolean quoted) {
 				var = getherevar(p);
 				if (var == NULL) {
 					freebuffer(buf);
-					p->input->ignoreeof = FALSE;
+					p->input->eof = FALSE;
 					return NULL;
 				}
 				*tailp = treecons(var, NULL);
@@ -88,8 +87,6 @@ extern Tree *snarfheredoc(Parser *p, const char *eof, Boolean quoted) {
 				break;
 		}
 	}
-
-	p->input->ignoreeof = FALSE;
 	return tree->CDR == NULL ? tree->CAR : tree;
 }
