+++ prim-etc.c
@@ -149,67 +149,16 @@ PRIM(var) {
 	return list;
 }
 
-PRIM(newparse) {
+PRIM(parse) {
 	List *result;
 	Ref(List *, reader, list);
 	Ref(Tree *, tree, NULL);
-	tree = parse(reader, NULL, NULL);
+	tree = parse(reader);
 	result = (tree == NULL)
 		   ? NULL
 		   : mklist(mkterm(NULL, mkclosure(gcmk(nThunk, tree), NULL)),
 			    NULL);
 	RefEnd2(tree, reader);
-	return result;
-}
-
-static void loginput(char *input) {
-	char *c;
-	List *fn = varlookup("fn-%write-history", NULL);
-	if (!isinteractive() || !isfromfd() || fn == NULL)
-		return;
-	for (c = input;; c++)
-		switch (*c) {
-		case '#': case '\n': return;
-		case ' ': case '\t': break;
-		default: goto writeit;
-		}
-writeit:
-	gcdisable();
-	Ref(List *, list, append(fn, mklist(mkstr(input), NULL)));
-	gcenable();
-	eval(list, NULL, 0);
-	RefEnd(list);
-}
-
-PRIM(parse) {
-	List *result;
-	Ref(char *, prompt1, NULL);
-	Ref(char *, prompt2, NULL);
-	Ref(List *, lp, list);
-	if (lp != NULL) {
-		prompt1 = getstr(lp->term);
-		if ((lp = lp->next) != NULL)
-			prompt2 = getstr(lp->term);
-	}
-	RefEnd(lp);
-	newhistbuffer();
-
-	Ref(Tree *, tree, NULL);
-	ExceptionHandler
-		tree = parse(NULL, prompt1, prompt2);
-	CatchException (ex)
-		Ref(List *, e, ex);
-		loginput(dumphistbuffer());
-		throw(e);
-		RefEnd(e);
-	EndExceptionHandler
-
-	loginput(dumphistbuffer());
-	result = (tree == NULL)
-		   ? NULL
-		   : mklist(mkterm(NULL, mkclosure(gcmk(nThunk, tree), NULL)),
-			    NULL);
-	RefEnd3(tree, prompt2, prompt1);
 	return result;
 }
 
@@ -316,65 +265,6 @@ PRIM(setmaxevaldepth) {
 	RefReturn(lp);
 }
 
-#if HAVE_READLINE
-PRIM(sethistory) {
-	if (list == NULL) {
-		sethistory(NULL);
-		return NULL;
-	}
-	Ref(List *, lp, list);
-	sethistory(getstr(lp->term));
-	RefReturn(lp);
-}
-
-PRIM(writehistory) {
-	if (list == NULL || list->next != NULL)
-		fail("$&writehistory", "usage: $&writehistory command");
-	loghistory(getstr(list->term));
-	return NULL;
-}
-
-PRIM(setmaxhistorylength) {
-	char *s;
-	int n;
-	if (list == NULL) {
-		setmaxhistorylength(-1); /* unlimited */
-		return NULL;
-	}
-	if (list->next != NULL)
-		fail("$&setmaxhistorylength", "usage: $&setmaxhistorylength [limit]");
-	Ref(List *, lp, list);
-	n = (int)strtol(getstr(lp->term), &s, 0);
-	if (n < 0 || (s != NULL && *s != '\0'))
-		fail("$&setmaxhistorylength", "max-history-length must be set to a positive integer");
-	setmaxhistorylength(n);
-	RefReturn(lp);
-}
-
-PRIM(resetterminal) {
-	resetterminal = TRUE;
-	return ltrue;
-}
-
-
-PRIM(readline) {
-	char *line;
-	char *prompt = (list == NULL ? "" : getstr(list->term));
-	if (list->next != NULL)
-		fail("$&readline", "usage: %read-line [prompt]");
-
-	do {
-		line = callreadline(prompt);
-	} while (line == NULL && errno == EINTR);
-
-	if (line == NULL)
-		return NULL;
-	list = mklist(mkstr(str("%s", line)), NULL);
-	efree(line);
-	return list;
-}
-#endif
-
 
 /*
  * initialization
@@ -392,7 +282,6 @@ extern Dict *initprims_etc(Dict *primdict) {
 	X(fsplit);
 	X(var);
 	X(parse);
-	X(newparse);
 	X(batchloop);
 	X(collect);
 	X(home);
@@ -404,12 +293,5 @@ extern Dict *initprims_etc(Dict *primdict) {
 	X(exitonfalse);
 	X(noreturn);
 	X(setmaxevaldepth);
-#if HAVE_READLINE
-	X(sethistory);
-	X(writehistory);
-	X(resetterminal);
-	X(setmaxhistorylength);
-	X(readline);
-#endif
 	return primdict;
 }
